+++ backend/src/controllers/incomeController.ts
@@ -41,7 +41,6 @@ export async function postUserIncome(ctx: Context) {
       date: body.date,
       description: body.description,
       amount: body.amount,
-      currency: body.currency,
       genre: body.genre,
     }).save();
     return ctx.json(newUserIncome);
