+++ frontend/src/components/ExpenseForm.tsx
@@ -26,9 +26,6 @@ const schema = z.object({
   genre: ExpenseGenre,
   amount: z.number(),
   date: z.date(),
-  currency: z.string()
-    .regex(/^[A-Z]{3}/, "must be 3 digit currency Code")
-    .default("JPY"),
 });
 
 export type ExpenseForm = z.infer<typeof schema>
@@ -49,7 +46,7 @@ export default function ExpenseForm(props: Props) {
     resolver: zodResolver(schema)
   })
 
-  const params =  useSearchParams().get("yearMonth");
+  const params = useSearchParams().get("yearMonth");
   const currentYearMonth = params
     ? DateTime.fromFormat(params, "yyyyMM").set({ day: DateTime.now().day }).toJSDate()
     : new Date();
@@ -82,28 +79,13 @@ export default function ExpenseForm(props: Props) {
       )}
       <form onSubmit={handleSubmit(onSubmit)}>
         <TextField
-          key={"description"}
-          id="description"
-          label="概要"
-          type="text"
-          placeholder="出費の概要"
-          defaultValue={null}
-          {...register("description")}
-          error={!!errors["description"]}
-          helperText={errors["description"]?.message}
-          variant="outlined"
-          margin="normal"
-          fullWidth
-        />
-
-        <TextField
           key={"amount"}
           id="amount"
           label="金額"
           type="number"
           placeholder="金額を数字で入力"
           defaultValue={null}
-          {...register('amount' , { valueAsNumber : true})}
+          {...register('amount', { valueAsNumber: true })}
           error={!!errors["amount"]}
           helperText={errors["amount"]?.message}
           variant="outlined"
@@ -131,6 +113,21 @@ export default function ExpenseForm(props: Props) {
             </MenuItem>
           ))}
         </TextField>
+
+        <TextField
+          key={"description"}
+          id="description"
+          label="概要"
+          type="text"
+          placeholder="出費の概要"
+          defaultValue={null}
+          {...register("description")}
+          error={!!errors["description"]}
+          helperText={errors["description"]?.message}
+          variant="outlined"
+          margin="normal"
+          fullWidth
+        />
         {/*TODO:　これを理解 */}
         <Controller
           name="date"
@@ -145,21 +142,6 @@ export default function ExpenseForm(props: Props) {
             />
           )}
         />
-
-        <TextField
-          key={"currency"}
-          id="currency"
-          label="通貨"
-          type="text"
-          placeholder="通貨を記入"
-          defaultValue={"JPY"}
-          {...register("currency")}
-          error={!!errors["currency"]}
-          helperText={errors["currency"]?.message}
-          variant="outlined"
-          margin="normal"
-          fullWidth
-        />
         <Button
           type="submit"
           variant="contained"
@@ -167,7 +149,7 @@ export default function ExpenseForm(props: Props) {
           sx={{ mt: 2 }}
           disabled={isSubmitting}
         >
-          {isSubmitting ? "submitting" : "SUBMIT"}
+          {isSubmitting ? "submitting...." : "SUBMIT"}
         </Button>
         <Button
           type="reset"
