+++ frontend/src/types/transaction.d.ts
@@ -4,7 +4,7 @@ type Transaction = {
   yearMonth: string;
   description: string;
   amount: number;
-  currency: string;
+  // currency: string; // not needed now
   genre: string;
   createdAt: string;
   updatedAt: string;
